# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a obs server.",
        "title": "obs Swagger API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/mid/limit": {
            "post": {
                "description": "update rate limit middleware config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "middleware"
                ],
                "summary": "update rate limit",
                "parameters": [
                    {
                        "description": "rate limit config",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dynamic.RateLimit"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/mid/switch": {
            "get": {
                "description": "list switch middleware config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "middleware"
                ],
                "summary": "list switch",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dynamic.Switch"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "update switch middleware config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "middleware"
                ],
                "summary": "update switch",
                "parameters": [
                    {
                        "description": "switch config",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dynamic.Switch"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/debug/pprof/heap": {
            "get": {
                "description": "get pprof heap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pprof"
                ],
                "summary": "pprof heap",
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/debug/pprof/index": {
            "get": {
                "description": "get pprof profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pprof"
                ],
                "summary": "pprof profile",
                "parameters": [
                    {
                        "type": "string",
                        "default": "30",
                        "description": "second default(30)",
                        "name": "seconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/debug/pprof/profile": {
            "get": {
                "description": "get pprof profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pprof"
                ],
                "summary": "pprof profile",
                "parameters": [
                    {
                        "type": "string",
                        "default": "30",
                        "description": "second default(30)",
                        "name": "seconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/debug/pprof/trace": {
            "get": {
                "description": "get pprof trace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "pprof"
                ],
                "summary": "pprof trace",
                "parameters": [
                    {
                        "type": "string",
                        "default": "30",
                        "description": "second default(30)",
                        "name": "seconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dynamic.BalanceNode": {
            "type": "object",
            "required": [
                "host"
            ],
            "properties": {
                "add": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "scheme": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dynamic.RateLimit": {
            "type": "object",
            "properties": {
                "burst": {
                    "type": "integer"
                },
                "every": {
                    "type": "string"
                }
            }
        },
        "dynamic.Switch": {
            "type": "object",
            "required": [
                "node",
                "service_name"
            ],
            "properties": {
                "add": {
                    "type": "boolean"
                },
                "node": {
                    "type": "object",
                    "$ref": "#/definitions/dynamic.BalanceNode"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}